cast is safe here because read_size is always >=0
Avoids a warning from MSVC

+++ include/osmium/io/detail/read_write.hpp
@@ -219,7 +219,7 @@ namespace osmium {
                         return false;
                     }
                     assert(read_size <= to_read);
-                    to_read -= read_size;
+                    to_read -= static_cast<unsigned int>(read_size);
                 }
 
                 return true;
